fixing tests

+++ src/builder.ts
@@ -90,6 +90,9 @@ export class DoculaBuilder {
 				`${siteRelativePath}/logo.svg`,
 				`${this.options.outputPath}/logo.svg`,
 			);
+		}
+
+		if (fs.existsSync(`${siteRelativePath}/logo_horizontal.png`)) {
 			await fs.promises.copyFile(
 				`${siteRelativePath}/logo_horizontal.png`,
 				`${this.options.outputPath}/logo_horizontal.png`,
